Revert "Update BC-004"
