+++ CarpetAndBeer/src/main/kotlin/shop/carpetandbeer/CarpetAndBeerApplication.kt
@@ -1,0 +1,11 @@
+package shop.carpetandbeer
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class CarpetAndBeerApplication
+
+fun main(args: Array<String>) {
+	runApplication<CarpetAndBeerApplication>(*args)
+}
