+++ CarpetAndBeer/src/main/kotlin/shop/carpetandbeer/repository/ItemRepository.kt
@@ -1,0 +1,7 @@
+package shop.carpetandbeer.repository
+
+import org.springframework.data.mongodb.repository.MongoRepository
+import shop.carpetandbeer.model.Item
+
+interface ItemRepository : MongoRepository<Item, String> {
+}
