+++ CarpetAndBeer/src/main/kotlin/shop/carpetandbeer/repository/UserRepository.kt
@@ -1,0 +1,7 @@
+package shop.carpetandbeer.repository
+
+import org.springframework.data.mongodb.repository.MongoRepository
+import shop.carpetandbeer.model.User
+
+interface UserRepository : MongoRepository<User, String> {
+}
