+++ CarpetAndBeer/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+spring.application.name=CarpetAndBeer
+server.port=8080
+server.address=0.0.0.0
+
+spring.data.mongodb.uri=mongodb+srv://integration:test@cluster0.kascvla.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
+spring.data.mongodb.auto-index-creation=true
+spring.data.mongodb.database=carpetandbeers
+
+logging.level.org.springframework.security=DEBUG
