+++ CarpetAndBeer/src/test/kotlin/shop/carpetandbeer/CarpetAndBeerApplicationTests.kt
@@ -1,0 +1,13 @@
+package shop.carpetandbeer
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class CarpetAndBeerApplicationTests {
+
+	@Test
+	fun contextLoads() {
+	}
+
+}
