--- src/frontend/eslint.config.mjs
@@ -1,5 +1,0 @@
-import antfu from '@antfu/eslint-config'
-
-export default antfu({
-  react: true,
-})
