+++ src/frontend/package.json
@@ -15,18 +15,13 @@
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
     "@fontsource/roboto": "^5.0.12",
-    "@fortawesome/fontawesome-free": "^6.5.2",
     "@mui/icons-material": "^5.15.15",
     "@mui/material": "^5.15.15",
-    "mdb-react-ui-kit": "^8.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.1.0",
     "react-material-ui-carousel": "^3.4.2",
-    "react-responsive-carousel": "^3.2.23",
-    "react-router-dom": "^6.23.0",
-    "zukeeper": "^1.0.2",
-    "zustand": "^4.5.2"
+    "react-router-dom": "^6.23.0"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^2.13.3",
