+++ src/frontend/src/App.tsx
@@ -3,9 +3,6 @@ import './App.css'
 import LandingPage from './pages/LandingPage'
 import { Route, Routes } from 'react-router-dom'
 import ContactPage from './pages/ContactPage'
-import ProfilePage from './pages/ProfilePage'
-import RegisterPage from './pages/RegisterPage'
-import LoginForm from './pages/LoginPage'
 
 const theme = createTheme({
   palette: {
@@ -35,9 +32,6 @@ function App() {
       <Routes>
         <Route path="/" element={<LandingPage />}></Route>
         <Route path="/contact" element={<ContactPage />}></Route>
-        <Route path="/profile" element={<ProfilePage />}></Route>
-        <Route path="/register" element={<RegisterPage />}></Route>
-        <Route path="/login" element={<LoginForm />}></Route>
       </Routes>
     </ThemeProvider>
   )
