--- src/frontend/src/models/User.ts
@@ -1,6 +1,0 @@
-export default interface User {
-  id: number
-  name: string
-  email: string
-  role: string
-}
