+++ src/frontend/src/pages/ContactPage.tsx
@@ -1,5 +1,4 @@
 import Navbar from '../components/navbar/Navbar'
-// import Footer from '../components/Footer'
 
 export default function ContactPage() {
   return (
