--- src/main/kotlin/com/beersAndCarpets/repository/ItemRepository.kt
@@ -1,8 +1,0 @@
-package com.beersAndCarpets.repository
-
-import com.beersAndCarpets.model.Item
-import org.springframework.data.mongodb.repository.MongoRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface ItemRepository: MongoRepository<Item, String>
