--- src/main/kotlin/com/beersAndCarpets/repository/UserRepository.kt
@@ -1,6 +1,0 @@
-package com.beersAndCarpets.repository
-
-import com.beersAndCarpets.model.User
-import org.springframework.data.mongodb.repository.MongoRepository
-
-interface UserRepository : MongoRepository<User, String>
