--- src/main/resources/application.properties
@@ -1,5 +1,0 @@
-spring.application.name=BeersAndCarpets
-server.port=8080
-
-spring.data.mongodb.uri=mongodb+srv://integration:testuser@cluster0.kascvla.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
-spring.data.mongodb.database=Cluster0
