--- src/test/kotlin/com/beersAndCarpets/BeersAndCarpetsApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.beersAndCarpets
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class BeersAndCarpetsApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
